Make the thread cpu time provider an instance variable. Fixes the spock test failures when the static provider is not recreated between tests.
